Separate test files for each inspector type
