+++ kotest-assertions/kotest-assertions-core/src/jvmTest/kotlin/com/sksamuel/kotest/inspectors/MapInspectorsTest.kt
@@ -24,25 +24,14 @@ import io.kotest.inspectors.forSome
 import io.kotest.inspectors.forSomeKeys
 import io.kotest.inspectors.forSomeValues
 import io.kotest.inspectors.forValuesExactly
-import io.kotest.matchers.comparables.beGreaterThan
-import io.kotest.matchers.comparables.beLessThan
 import io.kotest.matchers.ints.shouldBeGreaterThan
 import io.kotest.matchers.maps.shouldContain
-import io.kotest.matchers.should
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 
 @Suppress("ConstantConditionIf")
-class InspectorsTest : WordSpec() {
-
-   private val list = listOf(1, 2, 3, 4, 5)
-   private val array = arrayOf(1, 2, 3, 4, 5)
+class MapInspectorsTest : WordSpec() {
    private val map = mapOf(1 to "1", 2 to "2", 3 to "3", 4 to "4", 5 to "5")
-
-   data class DummyEntry(
-      val id: Int,
-      val name: String,
-   )
 
    init {
 
@@ -77,73 +66,17 @@ class InspectorsTest : WordSpec() {
       }
 
       "forAll" should {
-         "pass if all elements of an array pass" {
-            array.forAll {
-               it.shouldBeGreaterThan(0)
-            }
-         }
-         "pass if all elements of a collection pass" {
-            list.forAll {
-               it.shouldBeGreaterThan(0)
-            }
-         }
          "pass if all entries of a map pass" {
             map.forAll {
                it.key.shouldBe(it.value.toInt())
             }
          }
          "return itself" {
-            array.forAll {
-               it.shouldBeGreaterThan(0)
-            }.forAll {
-               it.shouldBeGreaterThan(0)
-            }
-
-            list.forAll {
-               it.shouldBeGreaterThan(0)
-            }.forAll {
-               it.shouldBeGreaterThan(0)
-            }
-
             map.forAll {
                it.key.shouldBe(it.value.toInt())
             }.forAll {
                it.key.shouldBe(it.value.toInt())
             }
-         }
-         "fail when an exception is thrown inside an array" {
-            shouldThrowAny {
-               array.forAll {
-                  if (true) throw NullPointerException()
-               }
-            }.message shouldBe "0 elements passed but expected 5\n" +
-               "\n" +
-               "The following elements passed:\n" +
-               "--none--\n" +
-               "\n" +
-               "The following elements failed:\n" +
-               "1 => java.lang.NullPointerException\n" +
-               "2 => java.lang.NullPointerException\n" +
-               "3 => java.lang.NullPointerException\n" +
-               "4 => java.lang.NullPointerException\n" +
-               "5 => java.lang.NullPointerException"
-         }
-         "fail when an exception is thrown inside a list" {
-            shouldThrowAny {
-               list.forAll {
-                  if (true) throw NullPointerException()
-               }
-            }.message shouldBe "0 elements passed but expected 5\n" +
-               "\n" +
-               "The following elements passed:\n" +
-               "--none--\n" +
-               "\n" +
-               "The following elements failed:\n" +
-               "1 => java.lang.NullPointerException\n" +
-               "2 => java.lang.NullPointerException\n" +
-               "3 => java.lang.NullPointerException\n" +
-               "4 => java.lang.NullPointerException\n" +
-               "5 => java.lang.NullPointerException"
          }
          "fail when an exception is thrown inside a map" {
             shouldThrowAny {
@@ -195,22 +128,6 @@ class InspectorsTest : WordSpec() {
       }
 
       "forNone" should {
-         "pass if no elements pass fn test for a list" {
-            list.forNone {
-               it shouldBe 10
-            }
-         }
-         "pass if no elements pass fn test for an array" {
-            array.forNone {
-               it shouldBe 10
-            }
-         }
-         "pass if an element throws an exception" {
-            val items = listOf(1, 2, 3)
-            items.forNone {
-               if (true) throw NullPointerException()
-            }
-         }
          "pass if no entries of a map pass" {
             map.forNone {
                it shouldBe mapOf(10 to "10").entries.first()
@@ -222,64 +139,10 @@ class InspectorsTest : WordSpec() {
             }
          }
          "return itself" {
-            list.forNone {
-               it shouldBe 10
-            }.forNone {
-               it shouldBe 10
-            }
-            array.forNone {
-               it shouldBe 10
-            }.forNone {
-               it shouldBe 10
-            }
             map.forNone {
                it shouldBe mapOf(10 to "10").entries.first()
-            }
-         }
-         "fail if one elements passes fn test" {
-            shouldThrow<AssertionError> {
-               list.forNone {
-                  it shouldBe 4
-               }
-            }.message shouldBe """1 elements passed but expected 0
-
-The following elements passed:
-4
-
-The following elements failed:
-1 => expected:<4> but was:<1>
-2 => expected:<4> but was:<2>
-3 => expected:<4> but was:<3>
-5 => expected:<4> but was:<5>"""
-         }
-         "fail if all elements pass fn test" {
-            shouldThrow<AssertionError> {
-               list.forNone {
-                  it should beGreaterThan(0)
-               }
-            }.message shouldBe """5 elements passed but expected 0
-
-The following elements passed:
-1
-2
-3
-4
-5
-
-The following elements failed:
---none--"""
-         }
-         "work inside assertSoftly block" {
-            val dummyEntries = listOf(
-               DummyEntry(id = 1, name = "first"),
-               DummyEntry(id = 2, name = "second"),
-            )
-
-            assertSoftly(dummyEntries) {
-               forNone {
-                  it.id shouldBe 3
-                  it.name shouldBe "third"
-               }
+            }.forNone {
+               it shouldBe mapOf(10 to "10").entries.first()
             }
          }
          "fail if one entry passes fn test" {
@@ -356,86 +219,16 @@ The following elements failed:
       }
 
       "forSome" should {
-         "pass if one elements pass test"  {
-            list.forSome {
-               it shouldBe 3
-            }
-         }
-         "pass if size-1 elements pass test"  {
-            list.forSome {
-               it should beGreaterThan(1)
-            }
-         }
          "pass if one entry pass test"  {
             map.forSome {
                it shouldBe mapOf(1 to "1").entries.first()
             }
          }
          "return itself" {
-            list.forSome {
-               it shouldBe 3
-            }.forSome {
-               it shouldBe 3
-            }
-
-            array.forSome {
-               it shouldBe 3
-            }.forSome {
-               it shouldBe 3
-            }
-
             map.forSome {
                it shouldBe mapOf(1 to "1").entries.first()
             }.forSome {
                it shouldBe mapOf(1 to "1").entries.first()
-            }
-         }
-         "fail if no elements pass test"  {
-            shouldThrow<AssertionError> {
-               array.forSome {
-                  it should beLessThan(0)
-               }
-            }.message shouldBe """No elements passed but expected at least one
-
-The following elements passed:
---none--
-
-The following elements failed:
-1 => 1 should be < 0
-2 => 2 should be < 0
-3 => 3 should be < 0
-4 => 4 should be < 0
-5 => 5 should be < 0"""
-         }
-         "fail if all elements pass test"  {
-            shouldThrow<AssertionError> {
-               list.forSome {
-                  it should beGreaterThan(0)
-               }
-            }.message shouldBe """All elements passed but expected < 5
-
-The following elements passed:
-1
-2
-3
-4
-5
-
-The following elements failed:
---none--"""
-         }
-         "work inside assertSoftly block" {
-            val dummyEntries = listOf(
-               DummyEntry(id = 1, name = "first"),
-               DummyEntry(id = 1, name = "first"),
-               DummyEntry(id = 2, name = "second"),
-            )
-
-            assertSoftly(dummyEntries) {
-               forSome {
-                  it.id shouldBe 1
-                  it.name shouldBe "first"
-               }
             }
          }
          "fail if no entries pass test"  {
@@ -513,79 +306,16 @@ The following elements failed:
       }
 
       "forOne" should {
-         "pass if one elements pass test"  {
-            list.forOne {
-               it shouldBe 3
-            }
-         }
          "pass if one entry pass test"  {
             map.forOne {
                it shouldBe mapOf(1 to "1").entries.first()
             }
          }
          "return itself" {
-            list.forOne {
-               it shouldBe 3
-            }.forOne {
-               it shouldBe 3
-            }
-
-            array.forOne {
-               it shouldBe 3
-            }.forOne {
-               it shouldBe 3
-            }
-
             map.forOne {
                it shouldBe mapOf(1 to "1").entries.first()
             }.forOne {
                it shouldBe mapOf(1 to "1").entries.first()
-            }
-         }
-         "fail if > 1 elements pass test"  {
-            shouldThrow<AssertionError> {
-               list.forOne {
-                  it should beGreaterThan(2)
-               }
-            }.message shouldBe """3 elements passed but expected 1
-
-The following elements passed:
-3
-4
-5
-
-The following elements failed:
-1 => 1 should be > 2
-2 => 2 should be > 2"""
-         }
-         "fail if no elements pass test"  {
-            shouldThrow<AssertionError> {
-               array.forOne {
-                  it shouldBe 22
-               }
-            }.message shouldBe """0 elements passed but expected 1
-
-The following elements passed:
---none--
-
-The following elements failed:
-1 => expected:<22> but was:<1>
-2 => expected:<22> but was:<2>
-3 => expected:<22> but was:<3>
-4 => expected:<22> but was:<4>
-5 => expected:<22> but was:<5>"""
-         }
-         "work inside assertSoftly block" {
-            val dummyEntries = listOf(
-               DummyEntry(id = 1, name = "first"),
-               DummyEntry(id = 2, name = "second"),
-            )
-
-            assertSoftly(dummyEntries) {
-               forOne {
-                  it.id shouldBe 1
-                  it.name shouldBe "first"
-               }
             }
          }
          "fail if > 1 entries pass test"  {
@@ -662,68 +392,16 @@ The following elements failed:
       }
 
       "forAny" should {
-         "pass if one elements pass test"  {
-            list.forAny {
-               it shouldBe 3
-            }
-         }
-         "pass if at least elements pass test"  {
-            list.forAny {
-               it should beGreaterThan(2)
-            }
-         }
          "pass if any entries pass test"  {
             map.forAny {
                mapOf(1 to "1", 2 to "2").shouldContain(it.toPair())
             }
          }
          "return itself" {
-            list.forAny {
-               it shouldBe 3
-            }.forAny {
-               it shouldBe 3
-            }
-
-            array.forAny {
-               it shouldBe 3
-            }.forAny {
-               it shouldBe 3
-            }
-
             map.forAny {
                mapOf(1 to "1", 2 to "2").shouldContain(it.toPair())
             }.forAny {
                mapOf(1 to "1", 2 to "2").shouldContain(it.toPair())
-            }
-         }
-         "fail if no elements pass test"  {
-            shouldThrow<AssertionError> {
-               array.forAny {
-                  it shouldBe 6
-               }
-            }.message shouldBe """0 elements passed but expected at least 1
-
-The following elements passed:
---none--
-
-The following elements failed:
-1 => expected:<6> but was:<1>
-2 => expected:<6> but was:<2>
-3 => expected:<6> but was:<3>
-4 => expected:<6> but was:<4>
-5 => expected:<6> but was:<5>"""
-         }
-         "work inside assertSoftly block" {
-            val dummyEntries = listOf(
-               DummyEntry(id = 1, name = "first"),
-               DummyEntry(id = 2, name = "second"),
-            )
-
-            assertSoftly(dummyEntries) {
-               forAny {
-                  it.id shouldBe 1
-                  it.name shouldBe "first"
-               }
             }
          }
          "fail if no entries pass test"  {
@@ -784,64 +462,10 @@ The following elements failed:
 
 
       "forExactly" should {
-         "pass if exactly k elements pass"  {
-            list.forExactly(2) {
-               it should beLessThan(3)
-            }
-         }
          "pass if exactly k entries pass"  {
             map.forExactly(4) {
                it shouldNotBe mapOf(1 to "1").entries.first()
             }
-         }
-         "fail if more elements pass test"  {
-            shouldThrow<AssertionError> {
-               list.forExactly(2) {
-                  it should beGreaterThan(2)
-               }
-            }.message shouldBe """3 elements passed but expected 2
-
-The following elements passed:
-3
-4
-5
-
-The following elements failed:
-1 => 1 should be > 2
-2 => 2 should be > 2"""
-         }
-         "fail if less elements pass test"  {
-            shouldThrow<AssertionError> {
-               array.forExactly(2) {
-                  it should beLessThan(2)
-               }
-            }.message shouldBe """1 elements passed but expected 2
-
-The following elements passed:
-1
-
-The following elements failed:
-2 => 2 should be < 2
-3 => 3 should be < 2
-4 => 4 should be < 2
-5 => 5 should be < 2"""
-         }
-         "fail if no elements pass test"  {
-            shouldThrow<AssertionError> {
-               array.forExactly(2) {
-                  it shouldBe 33
-               }
-            }.message shouldBe """0 elements passed but expected 2
-
-The following elements passed:
---none--
-
-The following elements failed:
-1 => expected:<33> but was:<1>
-2 => expected:<33> but was:<2>
-3 => expected:<33> but was:<3>
-4 => expected:<33> but was:<4>
-5 => expected:<33> but was:<5>"""
          }
          "fail if more entries pass test"  {
             shouldThrow<AssertionError> {
@@ -894,19 +518,27 @@ The following elements failed:
          }
       }
 
-      "forAtMostOnce" should {
-         "work inside assertSoftly block" {
-            val dummyEntries = listOf(
-               DummyEntry(id = 1, name = "first"),
-               DummyEntry(id = 2, name = "second"),
-            )
-
-            assertSoftly(dummyEntries) {
-               forAtMostOne {
-                  it.id shouldBe 1
-                  it.name shouldBe "first"
-               }
+      "forAtMostOne" should {
+         "pass if one elements pass test"  {
+            map.forAtMostOne {
+               it shouldBe mapOf(3 to "3")
             }
+         }
+         "fail if 2 elements pass test"  {
+            shouldThrow<AssertionError> {
+               map.forAtMostOne {
+                  mapOf(4 to "4", 5 to "5").shouldContain(it.toPair())
+               }
+            }.message shouldBe """2 elements passed but expected at most 1
+
+The following elements passed:
+4=4
+5=5
+
+The following elements failed:
+1=1 => Map should contain mapping 1=1 but was {4=4, 5=5}
+2=2 => Map should contain mapping 2=2 but was {4=4, 5=5}
+3=3 => Map should contain mapping 3=3 but was {4=4, 5=5}"""
          }
          "work inside assertSoftly block (for map)" {
             assertSoftly(map) {
@@ -919,18 +551,27 @@ The following elements failed:
       }
 
       "forAtLeastOne" should {
-         "work inside assertSoftly block" {
-            val dummyEntries = listOf(
-               DummyEntry(id = 1, name = "first"),
-               DummyEntry(id = 2, name = "second"),
-            )
-
-            assertSoftly(dummyEntries) {
-               forAtLeastOne {
-                  it.id shouldBe 1
-                  it.name shouldBe "first"
-               }
+         "pass if one elements pass test"  {
+            map.forAtLeastOne {
+               it shouldBe mapOf(3 to "3").entries.first()
             }
+         }
+         "fail if no elements pass test"  {
+            shouldThrow<AssertionError> {
+               map.forAtLeastOne {
+                  it shouldBe mapOf(22 to "22").entries.first()
+               }
+            }.message shouldBe """0 elements passed but expected at least 1
+
+The following elements passed:
+--none--
+
+The following elements failed:
+1=1 => expected:<22=22> but was:<1=1>
+2=2 => expected:<22=22> but was:<2=2>
+3=3 => expected:<22=22> but was:<3=3>
+4=4 => expected:<22=22> but was:<4=4>
+5=5 => expected:<22=22> but was:<5=5>"""
          }
          "work inside assertSoftly block (for map)" {
             assertSoftly(map) {
